Remove enum.h include of MarlinConfig.h
